Update CSP - moved to separate class, create Security sub folder
