+++ Core/Http/Response.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Core\Http;
 
+use Core\Security\ContentPolicy;
+
 readonly class Response
 {
     public function __construct(
@@ -16,8 +18,7 @@ readonly class Response
     public function send(): void
     {
         http_response_code($this->status);
-        // TODO: add option to specify custom security headers
-        header("Content-Security-Policy: default-src 'self'; font-src 'self' https://fonts.googleapis.com;");
+        ContentPolicy::use();
         if (!empty($this->headers)) {
             foreach ($this->headers as $header) {
                 header($header);
